pyproject.toml
[tool.poetry]
name = "mlds-mlops"
version = "0.1.0"
description = "MLOps study project for MLOps Course @ HSE MLDS Program"
authors = ["Maksim Ozerov <mvozerov@edu.hse.ru>"]
readme = "README.md"
packages = [{include = "mlds_mlops"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
dvc = "^3.38.1"
dvc-gdrive = "^3.0.0"
hydra-core = "^1.3.2"
scikit-learn = "^1.3.2"
skl2onnx = "^1.16.0"
onnxruntime = "^1.16.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[tool.black]
line-length = 99
target-version = ["py39"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 99
lines_after_imports = 2

[tool.flake8]
max-line-length = 99

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
